workers/audio_compress/requirements.txt
google-cloud-core==2.4.1
google-cloud-storage==2.17.0